JAMES-1877 Extract DNS resolution to a DNS helper
